Opprettet Service-klasser for alle Clients som ikke hadde det, og flyttet mapping inn i Servicene, så de ikke er i Route-en vår
