+++ src/main/kotlin/no/nav/medlemskap/services/aareg/AaRegDomene.kt
@@ -1,4 +1,4 @@
-package no.nav.medlemskap.modell.aareg
+package no.nav.medlemskap.services.aareg
 
 import java.time.LocalDate
 import java.time.LocalDateTime
