+++ src/main/kotlin/no/nav/medlemskap/services/inntekt/InntektMapper.kt
@@ -2,8 +2,6 @@ package no.nav.medlemskap.services.inntekt
 
 import no.nav.medlemskap.domene.Inntekt
 import no.nav.medlemskap.domene.InntektArbeidsforhold
-import no.nav.medlemskap.modell.inntekt.ArbeidsinntektMaaned
-import no.nav.medlemskap.modell.inntekt.InntektskomponentResponse
 
 //  Todo Bruker foreløpig ingen data fra inntekstkomponenten
 fun mapInntektResultat(@Suppress("UNUSED_PARAMETER") inntekt: InntektskomponentResponse): List<Inntekt> {
