+++ src/main/kotlin/no/nav/medlemskap/services/pdl/PdlDomene.kt
@@ -1,4 +1,4 @@
-package no.nav.medlemskap.modell.pdl
+package no.nav.medlemskap.services.pdl
 
 
 data class HentIdenterResponse(val data: Data, val errors: List<PdlError>?)
